update collect information about files uncovered by tests
